+++ src/fractal_task_tools/_args_schemas.py
@@ -13,11 +13,11 @@ from ._descriptions import _get_class_attrs_descriptions
 from ._descriptions import _get_function_args_descriptions
 from ._descriptions import _insert_class_attrs_descriptions
 from ._descriptions import _insert_function_args_descriptions
-from ._pydantic_generatejsonschema import CustomGenerateJsonSchema
 from ._signature_constraints import _extract_function
 from ._signature_constraints import _validate_function_signature
 from ._titles import _include_titles
 
+from ._generatejsonschema import CustomGenerateJsonSchema
 
 _Schema = dict[str, Any]